Move skip test into entry point test, remove unittest.mock
